+++ src/Mvc/Model/CronJob.php
@@ -7,51 +7,42 @@ $CronJob = [
 			'type'		=> 's',
 			'required'	=> true,
 			'size'		=> 255,
-			'date'		=> false,
 		],
 
 		'description' => [
 			'type'		=> 't',
 			'required'	=> false,
-			'date'		=> false,
 		],
 
 		'code' => [
 			'type'		=> 'lb',
 			'required'	=> false,
-			'date'		=> false,
-			'protected'	=> true,
 		],
 
 		// Schedule configuration
 		'schedule_type' => [
 			'type'		=> 's',
-			'required'	=> true,
+			'required'	=> false,
 			'size'		=> 20,
 			'default'	=> 'cron',
-			'date'		=> false,
 		],
 
 		'cron_expression' => [
 			'type'		=> 's',
 			'required'	=> false,
 			'size'		=> 100,
-			'date'		=> false,
 		],
 
 		'interval_seconds' => [
 			'type'		=> 'i',
 			'required'	=> false,
-			'size'		=> 11,
 			'unsigned'	=> true,
-			'date'		=> false,
 		],
 
 		'time_of_day' => [
 			'type'		=> 's',
 			'required'	=> false,
-			'size'		=> 8,
-			'date'		=> false,
+			'size'		=> 20,
 		],
 
 		'day_of_week' => [
@@ -59,7 +50,6 @@ $CronJob = [
 			'required'	=> false,
 			'size'		=> 1,
 			'unsigned'	=> true,
-			'date'		=> false,
 		],
 
 		'day_of_month' => [
@@ -67,7 +57,6 @@ $CronJob = [
 			'required'	=> false,
 			'size'		=> 2,
 			'unsigned'	=> true,
-			'date'		=> false,
 		],
 
 		'timezone' => [
@@ -75,7 +64,6 @@ $CronJob = [
 			'required'	=> false,
 			'size'		=> 50,
 			'default'	=> 'UTC',
-			'date'		=> false,
 		],
 
 		// Execution settings
@@ -85,16 +73,13 @@ $CronJob = [
 			'size'		=> 1,
 			'unsigned'	=> true,
 			'default'	=> 1,
-			'date'		=> false,
 		],
 
 		'timeout_seconds' => [
 			'type'		=> 'i',
 			'required'	=> false,
-			'size'		=> 11,
 			'unsigned'	=> true,
 			'default'	=> 300,
-			'date'		=> false,
 		],
 
 		'max_retries' => [
@@ -103,7 +88,6 @@ $CronJob = [
 			'size'		=> 1,
 			'unsigned'	=> true,
 			'default'	=> 3,
-			'date'		=> false,
 		],
 
 		'retry_strategy' => [
@@ -111,16 +95,13 @@ $CronJob = [
 			'required'	=> false,
 			'size'		=> 20,
 			'default'	=> 'exponential',
-			'date'		=> false,
 		],
 
 		'retry_delay_seconds' => [
 			'type'		=> 'i',
 			'required'	=> false,
-			'size'		=> 11,
 			'unsigned'	=> true,
 			'default'	=> 60,
-			'date'		=> false,
 		],
 
 		'allow_concurrent' => [
@@ -129,16 +110,13 @@ $CronJob = [
 			'size'		=> 1,
 			'unsigned'	=> true,
 			'default'	=> 0,
-			'date'		=> false,
 		],
 
 		// Dependencies
 		'depends_on_job' => [
 			'type'		=> 'i',
 			'required'	=> false,
-			'size'		=> 11,
 			'unsigned'	=> true,
-			'date'		=> false,
 			'fk'		=> [
 				'model'	=> 'CronJob',
 				'field'	=> 'id',
@@ -152,16 +130,13 @@ $CronJob = [
 			'size'		=> 1,
 			'unsigned'	=> true,
 			'default'	=> 0,
-			'date'		=> false,
 		],
 
 		'notify_after_failures' => [
 			'type'		=> 'i',
 			'required'	=> false,
-			'size'		=> 11,
 			'unsigned'	=> true,
 			'default'	=> 3,
-			'date'		=> false,
 		],
 
 		'notify_on_dead_letter' => [
@@ -170,22 +145,18 @@ $CronJob = [
 			'size'		=> 1,
 			'unsigned'	=> true,
 			'default'	=> 1,
-			'date'		=> false,
 		],
 
 		'slack_webhook' => [
 			'type'		=> 's',
 			'required'	=> false,
 			'size'		=> 512,
-			'date'		=> false,
-			'protected'	=> true,
 		],
 
 		'notification_email' => [
 			'type'		=> 's',
 			'required'	=> false,
 			'size'		=> 255,
-			'date'		=> false,
 		],
 
 		// Dead Letter Queue
@@ -195,36 +166,30 @@ $CronJob = [
 			'size'		=> 1,
 			'unsigned'	=> true,
 			'default'	=> 0,
-			'date'		=> false,
 		],
 
 		'dead_letter_reason' => [
 			'type'		=> 't',
 			'required'	=> false,
-			'date'		=> false,
 		],
 
 		'dead_letter_since' => [
 			'type'		=> 'i',
 			'required'	=> false,
-			'date'		=> true,
+			'unsigned'	=> true,
 		],
 
 		'consecutive_failures' => [
 			'type'		=> 'i',
 			'required'	=> false,
-			'size'		=> 11,
 			'unsigned'	=> true,
 			'default'	=> 0,
-			'date'		=> false,
 		],
 
 		// Context
 		'application' => [
 			'type'		=> 'i',
 			'required'	=> false,
-			'size'		=> 11,
-			'date'		=> false,
 			'fk'		=> [
 				'model'	=> 'Application',
 				'field'	=> 'id',
@@ -238,61 +203,16 @@ $CronJob = [
 			'size'		=> 1,
 			'unsigned'	=> true,
 			'default'	=> 0,
-			'date'		=> false,
 		],
 
 		'kyte_account' => [
 			'type'		=> 'i',
 			'required'	=> true,
-			'size'		=> 11,
 			'unsigned'	=> true,
-			'date'		=> false,
-		],
-
-		// Audit attributes
-		'created_by' => [
-			'type'		=> 'i',
-			'required'	=> false,
-			'date'		=> false,
-		],
-
-		'date_created' => [
-			'type'		=> 'i',
-			'required'	=> false,
-			'date'		=> true,
-		],
-
-		'modified_by' => [
-			'type'		=> 'i',
-			'required'	=> false,
-			'date'		=> false,
-		],
-
-		'date_modified' => [
-			'type'		=> 'i',
-			'required'	=> false,
-			'date'		=> true,
-		],
-
-		'deleted_by' => [
-			'type'		=> 'i',
-			'required'	=> false,
-			'date'		=> false,
-		],
-
-		'date_deleted' => [
-			'type'		=> 'i',
-			'required'	=> false,
-			'date'		=> true,
-		],
-
-		'deleted' => [
-			'type'		=> 'i',
-			'required'	=> false,
-			'size'		=> 1,
-			'unsigned'	=> true,
-			'default'	=> 0,
-			'date'		=> false,
+			'fk'		=> [
+				'model'	=> 'KyteAccount',
+				'field'	=> 'id',
+			],
 		],
 	],
 ];